fix bag with notification and cars

+++ ural/notification/views.py
@@ -18,7 +18,10 @@ def objectsNotify(cargs, cars):
     j=0
     while i < len(cargs) and j < len(cars):
         if cargs[i].id < cars[j].id:
-            objects.append(ObjectNotify(id=cars[j].id, car=cars[j]))
+            #print(cars[j].id)
+            obj = ObjectNotify(id=cars[j].id, car=cars[j])
+            objects.append(obj)
+            print(obj.id)
             j+=1
         else:
             objects.append(ObjectNotify(id=cargs[i].id, cargo=cargs[i]))
@@ -30,16 +33,18 @@ def objectsNotify(cargs, cars):
             i+=1
     else:
         while j < len(cars):
-            objects.append(ObjectNotify(id=cars[j], car=cars[j]))
+            objects.append(ObjectNotify(id=cars[j].id, car=cars[j]))
             j+=1
+
+
     return objects
 
 def arrayCars(cars):
     array=[]
     for i in cars:
-        car = VCar(i.car.id, i.car.car, i.car.capacity, i.car.volume, i.car.length, i.car.width, 
-                   i.car.height, i.car.where_from, i.car.where, i.car.ready_from, i.car.ready_to, 
-                   i.car.phone, i.car.comment,)
+        car = VCar(_id=i.car.id, _name=i.car.car, _capacity=i.car.capacity, _volume=i.car.volume, _length=i.car.length, _width=i.car.width, 
+                   _height=i.car.height, _where_from=i.car.where_from, _where=i.car.where, _ready_from=i.car.ready_from, _ready_to=i.car.ready_to, 
+                   _phone=i.car.phone, _comment=i.car.comment)
         typesBody = carTypeBody.objects.all().select_related('car', 'type_body')
         check = False
         j=0
@@ -62,6 +67,7 @@ def arrayCars(cars):
                 break    
             k+=1
         array.append(car)
+
     return array
 
 def arrayCargs(cargs):
@@ -78,6 +84,9 @@ def getObjects(notifications_cargo, notifications_car):
 # функция которая возврщает словарб являющийся контекстом для страницы
 def getContext(notifications_cargo, notifications_car, request):
     objects = getObjects(notifications_cargo, notifications_car)
+
+    for i in objects:
+        print(i.id)
 
     paginator = Paginator(objects, per_page=4)
     page_number = request.GET.get('page')
@@ -126,13 +135,14 @@ def my_responses(request):
             obj = notifyCargo.objects.get(id=obj_id)
             obj.delete()
         else:
-            obj_id = request.POST.get('obj_cargo_id')
-            obj = notifyCar.objects.get(id=request.POST.get('obj_car_id'))
-            obj.delete()
+            obj_id = request.POST.get('obj_car_id')
+            if notifyCar.objects.get(id=obj_id) != None:
+                obj = notifyCar.objects.get(id=obj_id)
+                obj.delete()
+
 
     notifications_cargo = notifyCargo.objects.all().filter(first_user=request.user).order_by("-id")
     notifications_car = notifyCar.objects.all().filter(first_user=request.user).order_by("-id")
-    
     context = getContext(notifications_cargo, notifications_car, request)
     return render(request, 'MyResponses.html', context=context)
